Fourth lesson

+++ src/main/java/uz/brogrammers/clinic/user/model/dto/response/UserResponse.java
@@ -2,7 +2,6 @@ package uz.brogrammers.clinic.user.model.dto.response;
 
 import lombok.Builder;
 import lombok.Data;
-import uz.brogrammers.clinic.user.model.entity.Role;
 import uz.brogrammers.clinic.user.model.enums.RoleName;
 
 import java.util.Set;
@@ -15,7 +14,7 @@ public class UserResponse {
     private String firstName;
     private String lastName;
     private String phone;
-    private Long departmentId;
+    private Integer departmentId;
     private Set<RoleName> roles;
     private boolean isAdmin;
